+++ src/text_shaper/open_shaper.cpp
@@ -472,6 +472,10 @@ namespace // {{{ helper
                                             static_cast<FT_UInt>(_dpi.y)); ec != FT_Err_Ok)
             {
                 debuglog(FontLoaderTag).write("Failed to FT_Set_Char_Size(size={}, dpi {}, source {}): {}\n", size, _dpi, _source, ftErrorStr(ec));
+                // If we cannot set the char-size, this font is most likely unusable for us.
+                // Specifically PCF files fail here and I do not know how to deal with them in that
+                // case, so do not use this font file at all.
+                return nullopt;
             }
         }
 
@@ -779,7 +783,7 @@ void open_shaper::shape(font_key _font,
 
     // Reshape each cluster individually.
     _result.clear();
-    int cluster = _clusters[0];
+    auto cluster = _clusters[0];
     int start = 0;
     for (int i = 1; i < _clusters.size(); ++i)
     {
@@ -796,13 +800,13 @@ void open_shaper::shape(font_key _font,
         }
     }
 
+    // shape last cluster
     auto const end = _clusters.size();
-    if (d->tryShapeWithFallback(_font, fontInfo, hbBuf, hbFont,
+    d->tryShapeWithFallback(_font, fontInfo, hbBuf, hbFont,
                                 _script, _presentation,
                                 _codepoints.substr(start, end - start),
                                 _clusters.subspan(start, end - start),
-                                _result))
-        return;
+                                _result);
 
     // last resort
     replaceMissingGlyphs(fontInfo.ftFace.get(), _result);